requirements.txt
pandas
numpy
sqlalchemy
scikit-learn
matplotlib
seaborn
streamlit
prophet
plotly
joblib
psycopg2-binary
requests
beautifulsoup4
selenium
camelot-py[cv]
prefect>=2.13.0
streamlit
PyPDF2
openpyxl
